+++ src/main/java/roomescape/controller/HomePageController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class HomeController {
+public class HomePageController {
 
     @GetMapping("/")
     public String homePage() {
