+++ src/main/java/roomescape/controller/theme/dto/CreateThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.theme.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
