+++ src/main/java/roomescape/controller/theme/dto/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.theme.dto;
 
 import java.util.List;
 import roomescape.model.Theme;
